Update SkTDArray::reserve to use int
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/600438
Commit-Queue: Herb Derby <herb@google.com>

+++ include/private/SkTDArray.h
@@ -188,7 +188,7 @@ public:
         fStorage.resize(count);
     }
 
-    void reserve(size_t n) {
+    void reserve(int n) {
         fStorage.reserve(n);
     }
